Initial move of the configuration-wizard to a new package
