+++ packages/configuration-wizard/Header.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import Icon from "../Plugin/Shared/components/Icon";
+import Icon from "yoast-components/composites/Plugin/Shared/components/Icon";
 
 /**
  * Onboarding wizard header.
