+++ packages/configuration-wizard/LoadingIndicator.js
