+++ packages/configuration-wizard/ProgressIndicator.js
