+++ packages/configuration-wizard/components/custom_components/ConnectGoogleSearchConsole.js
