+++ packages/configuration-wizard/components/custom_components/MailchimpSignup.js
