+++ packages/configuration-wizard/components/custom_components/PostTypeVisibility.js
