+++ packages/configuration-wizard/components/custom_components/PublishingEntity.js
@@ -1,4 +1,3 @@
-
 import React from "react";
 import PropTypes from "prop-types";
