+++ packages/configuration-wizard/config/production-config.js
