+++ packages/configuration-wizard/config/test-configuration.js
