+++ packages/configuration-wizard/helpers/ajaxHelper.js
