+++ packages/configuration-wizard/helpers/htmlDecoder.js
