+++ packages/configuration-wizard/jest/setupTests.js
@@ -1,0 +1,1 @@
+import "jest-styled-components";
