+++ packages/configuration-wizard/onboarding-wizard.scss
@@ -1,4 +1,5 @@
-@import "../../style-guide/colors";
+//TODO: Remove this reference.
+@import "../yoast-components/style-guide/colors";
 
 $prefix: "yoast-wizard";
