+++ packages/configuration-wizard/package.json
@@ -1,0 +1,66 @@
+{
+  "name": "@yoast/configuration-wizard",
+  "version": "1.0.0",
+  "description": "Package for our configuration wizard",
+  "main": "index.js",
+  "repository": "https://github.com/Yoast/javascript",
+  "author": "Yoast",
+  "license": "GPL",
+  "private": false,
+  "scripts": {
+    "test": "jest"
+  },
+  "dependencies": {
+    "yoast-components": "^4.22.1",
+    "prop-types": "^15.7.2",
+    "interpolate-components": "^1.1.1",
+    "lodash": "^4.17.11",
+    "whatwg-fetch": "1.1.1"
+  },
+  "devDependencies": {
+    "jest": "^22.4.3",
+    "jest-styled-components": "^5.0.1",
+    "babel-core": "^6.25.0",
+    "babel-jest": "^22.4.3",
+    "babel-loader": "^7.1.1",
+    "babel-plugin-styled-components": "^1.5.1",
+    "babel-polyfill": "^6.23.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "enzyme": "3.9.0",
+    "enzyme-adapter-react-16": "1.10.0",
+    "enzyme-to-json": "3.3.5",
+    "react": "16.6.3",
+    "react-dom": "16.6.3"
+  },
+  "peerDependencies": {
+    "material-ui": "^0.18.6",
+    "react": "^16.2.0",
+    "react-dom": "16.2.0",
+    "react-intl": "^2.4.0"
+  },
+  "jest": {
+    "testURL": "http://localhost",
+    "testRegex": ".*Test.js$",
+    "setupTestFrameworkScriptFile": "<rootDir>/jest/setupTests.js",
+    "collectCoverageFrom": [
+      "**/*.js",
+      "!**/*.*Test.js",
+      "!grunt/**/*.js",
+      "!*.js",
+      "!app/**/*.js",
+      "!a11y/examples/example.js",
+      "!utils/helpers/*.js",
+      "!coverage/**/*.js"
+    ],
+    "snapshotSerializers": [
+      "enzyme-to-json/serializer"
+    ],
+    "transform": {
+      "^.+\\.jsx?$": "babel-jest"
+    },
+    "transformIgnorePatterns": [
+      "/node_modules/(?!yoastseo|lodash-es).+\\.js$"
+    ]
+  }
+}
