+++ packages/yoast-components/css/all.scss
@@ -1,4 +1,4 @@
 @import "accessibility";
-@import "../composites/OnboardingWizard/onboarding-wizard";
+@import "yoast-configuration-wizard/onboarding-wizard";
 @import "../composites/basic/loader";
 @import "loadingSpinner";
